don't treat timers separately
This is ancient code, from a time in which we had to handle the preempt
timer in a special way. Our uring code is now more capable and more
generic, and the cancellation queue allow us to handle the preempt timer
like any other event.

I don't know for sure if this is the cause of Issue 326, but I suspect
it is: if a timer fires when we are processing the cancellations, and it
happens to be the only event that fires, it is theoretically possible
that we go to sleep, in which case we'll never wake up.

Although I am not sure this is the cause of the bug, it is one fewer
special case.

Ref #326